Transitioned to PostgreSQL!
This is insanely exciting. My. First. Database!

I used Peewee, a Postgres wrapper for Python that lets me build the
data model out of regular-ish Python classes. It’s great.

And while I could’ve gone with the built in SQLite, I decided on
Postgres because it’ll be great for transitioning to a web app,
eventually. Right now I’m planning on using Flask for that, because
it’s lightweight & lets me bring my own database.

This is probably a good stopping point for tonight.