Data helper is now smart enough to not duplicate schools or programs
Cleaned data helper — it doesn’t duplicate schools or programs anymore,
but is smart enough to be able to update info that’s changed in
pre-existing records.

This is good stuff to build on. Next step is figuring out what exactly
I want to import from IPEDS data, then expanding the data model &
import functions to do just that.