Done refactoring the import functions!
It still needs testing, but this is great: I’ve successfully
refactored/restructured/reworked the functions for importing data.

It’ll now be *much* easier to expand the data models, without having to
add much more code to the data_helper.

Next step: testing. Then we move on to expanding the data model!