Began work on the solver!
Today I began work on the next big component of the project: the
solver. The solver will include functions to derive ROI and find the
best programs for a student, and, eventually, put together entire
pathways.

I had to do a bit more building out of the data model to pull this off,
which included some changes to data_helper.