Added convenience functions
Added convenience functions to pathways and their steps — made it a
little easier to get info about the steps, and much easier to get a
pathway’s steps sorted in the proper order.

There still probably is a better fix for this — some way to get the
database to return records in the order I want — but this is nice for
now.