Can now make multiple pathways per template
Adjusted the make pathways function to take a “how_many” variable, then
return that many pathways per template (as long as there are enough
schools in the database to do so).

It’s a simple implementation — right now it will never use the same
school in two different pathways, and that might be worth changing. But
it’ll work for now!