Demo.py can manage schools data
Added a hidden menu (accessed by typing “hidden”) that handles
importing and deleting school data. It’s a blunt tool, but good to have.

Also fixed a weird bug in generating pathways. But I’m probably going
to have to refactor that part of the solver in the future; it’s messy.