Added the Report table to the database
Pathways now belong to a Report, which belongs to a Student. Reports
now track careers instead of students — this future-proofs us for a
time when students can generate many reports for different careers.

I’ve adjusted everything to work with this new setup, but
text_interface.py is still a bit janky in the “delete_pathways”
function.