User authentication is working!
This is a big moment — in terms of functionality, everything is done
except for the employees’ ability to edit reports.

Obviously there’s tons of work to be done in terms of how the website
is formatted and hyperlinked together. But this is still a pretty great
place to get to.