Everything SEEMS to be restored...
Okay, fingers crossed that everything is back to the way it was.

I just ran some tests, created students, a couple customer users, and
an employee user. I had to recreate some missing templates, but it all
seems to be good now.

God, I hope this doesn’t mess anything up. If necessary, backups are in
/HiUni Reports Backup/