Added a navigation bar, and more sane redirects
It’s super simple, but the site is now actually navigable: there’s a
bar across the top with links to everywhere you’d need to go, and smart
redirects for cases where a user only has one student or a student only
has one report.

Next: I’d like to add the ability to make multiple reports for a
student. Why not?