Feature complete for the demo!
I’m feature complete for the demo!

If I have time this weekend, I still wouldn’t mind adding the ability
to generate multiple reports per student — it seems like a piece of
functionality that Howard would really love.

And: it’d be nice to present slightly different UI depending on if the
customer is an organization or not. I’d like to do this by just
detecting whether the “organization name” attribute is null or not.
This might involve some javascript on the registration page.

But otherwise we’re just in UI cleaning territory. Which is a great
place to be!