Cleaned up registering new users, separated customers and employees
Employees can now only be registered by a pre-existing employee who’s
logged in.

If no employees exist, and if the app is in debug mode, restarting
app.py will generate an admin account to use to register employees. I
should make sure to disable this (or just disable debug mode) before
opening this thing to the public.

Customer registration is cleaner now, and allows customers to choose to
register as an organization.