Removed PointField attempt
When I was researching using PostGIS, I started trying to figure out
how to add a custom field to Peewee’s model fields. Removed it now,
because I’m going to try using earthdistance instead of PostGIS, but
this is still interesting stuff.

Also: I’m going to look into overriding the save() function on the
student model to automatically save the latitude and longitude based on
the student’s city & state (but only if latitude and longitude are
null).