Almost there on image compression
Longterm, I need to store student’s photos in S3 and reference them.

For now I’m just trying to get compressing images and saving them to
the database in a bytea column to work.

I believe the next step is saving to a temp file, then grabbing that
file and putting it in the database.