Compressing images is working
Here’s the lame hack I had to settle on: saving a temporary image file
so I can open it back up and read it.

This appears to be a common problem, and there should be some way to
use TemporaryFile() to get around it, but I couldn’t find a way to
write a PIL Image to TemporaryFile().