Removed sidebar.html
Removed the sidebar from our UI entirely! We’ll use the navbar and
dropdowns for everything now — it’s simpler and more mobile-friendly!

I have a lot of changes to make to the backend — I need to store images
in S3, for starters, but I also need to expand my database to store
separate first and last names for students, individuated descriptions
for pathway steps, and pathway taglines.