Adding S3 storage for career banner images
It’s a work in progress. Currently all the code for getting a
connection to S3 and uploading a file lives in the add_career() view,
but I’ll move it to an async function in data_helper once I’m sure it’s
working.

And if this pans out, I could switch student photos to S3!