A good place!
Okay! We’re in a good place. Successfully uploading files to S3, but
not before renaming them to something appropriate, and with appropriate
permissions.

I’d like to be returning full-fledged files from the image property,
which right now only return urls to the files. But that’s for another
day.